Seperated reservation button
